pyproject.toml
[project]
name = "afl-fantasy-agent"
version = "0.1.0"
description = "Multi-agent AFL Fantasy Football intelligence system"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "supabase>=2.10.0",
    "apscheduler>=3.10.4",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "playwright>=1.44.0",
]

[project.scripts]
afl-agent = "afl_fantasy.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["afl_fantasy"]
